proje

+++ rentacar/src/main/java/com/visionrent/service/CarService.java
@@ -1,6 +1,4 @@
-package com.visionrent.security.service;
-
-
+package com.visionrent.service;
 
 import java.util.HashSet;
 import java.util.List;
@@ -8,11 +6,11 @@ import java.util.Set;
 import java.util.function.Function;
 
 
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+
 import com.visionrent.domain.Car;
 import com.visionrent.domain.ImageFile;
 import com.visionrent.dto.CarDTO;
@@ -22,82 +20,88 @@ import com.visionrent.exception.ResourceNotFoundException;
 import com.visionrent.exception.message.ErrorMessage;
 import com.visionrent.mapper.CarMapper;
 import com.visionrent.repository.CarRepository;
-import com.visionrent.service.ImageFileService;
+
 @Service
 public class CarService {
-	
+
 	@Autowired
 	private CarRepository carRepository;
-	
+
 	@Autowired
-	private ImageFileService imageFileService ;
-	
+	private ImageFileService imageFileService;
+
 	@Autowired
 	private CarMapper carMapper;
-	public void saveCar(String imageId, CarDTO carDTO) {
-		
-		// image ID imageRepo da var mı ?
-		ImageFile imageFile =  imageFileService.findImageById(imageId);
-		// imageId daha önce başka bir araç ile eşleşmiş mi
-		    Integer usedCarCount =  carRepository.findCarCountByImageId(imageFile.getId());
-		   
-		    if(usedCarCount>0) {
-		    	throw new ConflictException(ErrorMessage.IMAGE_USED_MESSAGE);
-		    }
-		   
-		    // mapper işlemi
-		    Car car =  carMapper.carDTOToCar(carDTO);
-		   
-		    Set<ImageFile> imFiles = new HashSet<>();
-		    imFiles.add(imageFile);
-		   
-		    car.setImage(imFiles);
-		   
-		    carRepository.save(car);
-	}
-	//********************************
 	
-	public List<CarDTO> getAllCars() {
-		
-	List<Car>	carList = carRepository .findAll();
-    return carMapper.map(carList);
-		
-		
+	@Autowired
+	private ReservationService reservationService;
+
+	// *************************************
+
+	public void saveCar(String imageId, CarDTO carDTO) {
+
+		// image ID imageRepo da var mı ?
+		ImageFile imageFile = imageFileService.findImageById(imageId);
+		// imageId daha önce başka bir araç ile eşleşmiş mi
+		Integer usedCarCount = carRepository.findCarCountByImageId(imageFile.getId());
+
+		if (usedCarCount > 0) {
+			throw new ConflictException(ErrorMessage.IMAGE_USED_MESSAGE);
+		}
+
+		// mapper işlemi
+		Car car = carMapper.carDTOToCar(carDTO);
+
+		Set<ImageFile> imFiles = new HashSet<>();
+		imFiles.add(imageFile);
+
+		car.setImage(imFiles);
+
+		carRepository.save(car);
+
 	}
+
+	// ********************************
+
+	public List<CarDTO> getAllCars() {
+
+		List<Car> carList = carRepository.findAll();
+		return carMapper.map(carList);
+
+	}
+
+	// ***********************************
+
 	public Page<CarDTO> findAllWithPage(Pageable pageable) {
-		
-		Page<Car> carPage  = carRepository.findAll(pageable);
-		Page<CarDTO> carPageDTO = carPage.map(new Function<Car,CarDTO>(){
+
+		Page<Car> carPage = carRepository.findAll(pageable);
+		Page<CarDTO> carPageDTO = carPage.map(new Function<Car, CarDTO>() {
 			@Override
 			public CarDTO apply(Car car) {
 				return carMapper.carToCarDTO(car);
 			}
 		});
-		return carPageDTO;		
+		return carPageDTO;
 	}
-	
-	
-	//***********************************************************************
 
+	// *************************************
 	public CarDTO findById(Long id) {
-		
-		Car car = getCar(id);
-		
-		return carMapper.carToCarDTO(car);
-		
-		
-		
-		
-	}
-	
-	public Car getCar(Long id) {
-	Car car	=  carRepository.findCarById(id).orElseThrow(()->new
-				ResourceNotFoundException(String.format(ErrorMessage.RESOURCE_NOT_FOUND_MESSAGE, id)));
-	return car;
-	}
-	//***********************************************************************************
 
-public void updateCar(Long id, String imageId, CarDTO carDTO) {
+		Car car = getCar(id);
+
+		return carMapper.carToCarDTO(car);
+
+	}
+
+	public Car getCar(Long id) {
+		Car car = carRepository.findCarById(id).orElseThrow(
+				() -> new ResourceNotFoundException(String.format(ErrorMessage.RESOURCE_NOT_FOUND_MESSAGE, id)));
+		return car;
+	}
+
+	// *********************************************************************************
+
+	public void updateCar(Long id, String imageId, CarDTO carDTO) {
 		
 		Car car = getCar(id);
 		
@@ -134,9 +138,11 @@ public void updateCar(Long id, String imageId, CarDTO carDTO) {
 		car.getImage().add(imageFile);
 		
 		carRepository.save(car);
+
 	}
 
-//******************** DELETE**********************
+	//******************** DELETE**********************
+	
 	public void removeById(Long id) {
 		
 		Car car = getCar(id);
@@ -145,9 +151,56 @@ public void updateCar(Long id, String imageId, CarDTO carDTO) {
 			throw new BadRequestException(ErrorMessage.NOT_PERMITTED_METHOD_MESSAGE);
 		}
 		
+		 boolean exist =  reservationService.existsByCar(car);
+		 
+		 // reservasyon kontrolü
+		 if(exist) {
+			 throw new BadRequestException(ErrorMessage.CAR_USED_BY_RESERVATION_MESSAGE);
+		 }
+		
+		
 		carRepository.delete(car);
 	
-	
+		
 	}
 	
+	// ---> EKLENDİ
+	public Car getCarById(Long id) {
+		Car car= carRepository.findById(id).orElseThrow(()->new 
+				ResourceNotFoundException(String.format(ErrorMessage.RESOURCE_NOT_FOUND_MESSAGE, id)));
+		return car;
+	}
+
+	public List<Car> getAllCar() {
+		return carRepository.getAllBy();
+	}
+	
+
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+
 }
